back/naep/pyproject.toml
[project]
name = "naep"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas",email = "lucaspework@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14, <4.0"
dependencies = [
    "fastapi[standard] (>=0.122.0,<0.123.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.6,<0.15.0)"
]



[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev naep/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=naep -vv'
post_test = 'coverage html'